Finalização do primeiro curso

+++ Aula10/ex11.py
@@ -1,6 +1,6 @@
-rt1 = int(input('Digite o valor da primeira reta: '))
-rt2 = int(input('Digite o valor da segunda reta: '))
-rt3 = int(input('Digite o valor da terceira reta: '))
+rt1 = float(input('Digite o valor da primeira reta: '))
+rt2 = float(input('Digite o valor da segunda reta: '))
+rt3 = float(input('Digite o valor da terceira reta: '))
 if rt1 + rt2 > rt3:
     print('Podem formar um triângulo!')
 else:
